+++ src/app/router/index.ts
@@ -2,5 +2,6 @@ export * from './Route';
 export * from './RouteAdmin';
 export * from './RoutePublic';
 export * from './RoutePublicOnly';
+export * from './usePaginationFromUrl';
 export * from './useRedirectFromUrl';
 export * from './useSearchParams';
