+++ src/components/Pagination/index.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useContext, FC } from 'react';
+import React, { useContext, FC } from 'react';
 
 import {
   Box,
@@ -14,21 +14,7 @@ import {
   FiChevronRight,
 } from 'react-icons/fi';
 
-import { useSearchParams } from '@/app/router';
 import { Icon } from '@/components';
-
-export const usePaginationFromUrl = () => {
-  const { searchParams, setSearchParam } = useSearchParams();
-  const page = +(searchParams.get('page') ?? 1);
-  const setPage = useCallback(
-    (p) => {
-      const newPage = Math.max(1, p);
-      setSearchParam('page', `${newPage}`);
-    },
-    [setSearchParam]
-  );
-  return { page, setPage };
-};
 
 export const getPaginationInfo = ({
   page = 1,
